+++ prefect_gcp/__init__.py
@@ -11,6 +11,7 @@ from .credentials import GcpCredentials  # noqa
 from .secret_manager import GcpSecret  # noqa
 from .workers.vertex import VertexAIWorker  # noqa
 from .workers.cloud_run import CloudRunWorker  # noqa
+from .workers.cloud_run_v2 import CloudRunWorkerV2  # noqa
 
 register_renamed_module(
     "prefect_gcp.projects", "prefect_gcp.deployments", start_date="Jun 2023"